refractoring

+++ others/change.php
@@ -43,44 +43,57 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
 
             $mysqli = new mysqli($mysql_dbhost, $mysql_dbuser, "", $mysql_dbname);
 
-            if ($mysqli->connect_errno) {
-                printf("Cannot connect to mysql: %s\n", $mysqli->connect_error);
-                exit();
-            }
+            $sql = 'select id FROM goods WHERE id = ' . $id_need_set;
+            $result = $mysqli->query($sql);
 
-            $sql = "UPDATE goods SET id = '{$id_need_set}', title = '{$title}', 
+            if ($result->fetch_array(MYSQL_NUM)[0] != null) {
+                echo "<script>alert('id exist');</script>";
+            }
+            else {
+
+                if ($mysqli->connect_errno) {
+                    printf("Cannot connect to mysql: %s\n", $mysqli->connect_error);
+                    exit();
+                }
+
+                $sql = "UPDATE goods SET id = '{$id_need_set}', title = '{$title}', 
                     description = '{$description}', cost = '${cost}', url_image = '{$url_image}' 
                     WHERE id=" . $id_original;
 
-            $result = $mysqli->query($sql);
+                $result = $mysqli->query($sql);
 
-            if (!$result) {
-                die('Could not change data: ' . $mysqli->error);
+                if (!$result) {
+                    die('Could not change data: ' . $mysqli->error);
+                }
+
+                $new_arr = array("cost" => $cost, "description" => $description, "title" => $title,
+                    "url_image" => $url_image);
+
+                $memcache->delete($id_original);
+
+                $memcache->set($id_need_set, $new_arr);
+
+                update_chunk_delete($memcache, $id_original);
+
+                //print_r($memcache->get("ids_reversed_cost_100"));
+
+                update_chunk_create($memcache, $id_need_set, $cost);
+
+                unset($new_arr);
+
+                echo "<script>alert('done');</script>";
+
+
             }
 
-            $new_arr = array("cost" => $cost, "description" => $description, "title" => $title,
-                "url_image" => $url_image);
-
-            $memcache->delete($id_original);
-
-            $memcache->set($id_need_set, $new_arr);
-
-            update_chunk_delete($memcache, $id_original);
-            update_chunk_create($memcache, $id_need_set, $cost);
-
-            unset($new_arr);
-
-            echo "<script>alert('done');</script>";
-
             $mysqli->close();
-
             break;
 
         default:
             break;
     }
 
-    $memcache->close();
+
 }
 
 ?>
